+++ packages/react-inline-edit-extension/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.3.35](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.34...@patternfly/react-inline-edit-extension@4.3.35) (2020-05-14)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
+## [4.3.34](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.33...@patternfly/react-inline-edit-extension@4.3.34) (2020-05-13)
+
+**Note:** Version bump only for package @patternfly/react-inline-edit-extension
+
+
+
+
+
 ## [4.3.33](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-inline-edit-extension@4.3.32...@patternfly/react-inline-edit-extension@4.3.33) (2020-05-12)
 
 **Note:** Version bump only for package @patternfly/react-inline-edit-extension